refactor: passando ListProps para arquivo separado
