+++ frontend/src/components/List/List.tsx
@@ -1,20 +1,7 @@
 ﻿import React, { ReactNode } from 'react';
 import './List.css';
 
-export interface ListProps<T> {
-  items: T[];
-  renderItem?: (item: T, index: number) => ReactNode;
-  keyExtractor?: (item: T, index: number) => string;
-  onItemClick?: (item: T, index: number) => void;
-  className?: string;
-  style?: React.CSSProperties;
-  emptyMessage?: string;
-  loading?: boolean;
-  loadingComponent?: ReactNode;
-  divider?: boolean;
-  scrollable?: boolean;
-  maxHeight?: string;
-}
+import { ListProps } from './List.types';
 
 /**
  * Componente List reutilizável para exibir dados de forma dinâmica e personalizável
